fix: analysis graph needs to consult graph_cache for external sboms, added tests
