+++ etc/test-data/cyclonedx/rh/image_index_variants/example_container_variant_arm64.json
