+++ etc/test-data/cyclonedx/rh/rpm_src_binary/example_rpm_binaries.json
