+++ etc/test-data/cyclonedx/rh/upstream/example_upstream.json
